+++ packages/at_onboarding_cli/lib/src/util/create_at_client_cli.dart
@@ -33,16 +33,14 @@ Future<AtClient> createAtClient(
       (atKeysFilePath ?? '$homeDir/.atsign/keys/${atSign}_key.atKeys')
           .replaceAll('/', Platform.pathSeparator);
   String? localStoragePathToUse = storageDir?.path;
-  String? commitLogStoragePathToUse =
-      ('${storageDir?.path}/commit').replaceAll('/', Platform.pathSeparator);
   String downloadPathToUse = ('$homeDir!/.atsign/downloads/$atSign/$nameSpace')
       .replaceAll('/', Platform.pathSeparator);
 
   String rootServer = rootDomain ?? AuthCliArgs.defaultAtDirectoryFqdn;
-  
+
   // Parse rootServer using AtRootDomain
   AtRootDomain parsedRootDomain = AtRootDomain.parse(rootServer);
-  
+
   AtOnboardingPreference atOnboardingPreference = AtOnboardingPreference()
     ..atKeysFilePath = atKeysFilePathToUse
     ..namespace = nameSpace
@@ -50,7 +48,6 @@ Future<AtClient> createAtClient(
     ..rootPort = parsedRootDomain.rootPort
     ..passPhrase = passPhrase
     ..hiveStoragePath = localStoragePathToUse
-    ..commitLogPath = commitLogStoragePathToUse
     ..downloadPath = downloadPathToUse;
 
   AtOnboardingService atOnboardingService = AtOnboardingServiceImpl(